RECAP-673 - Modification done to update the updated data and insert into change log
